requirements-server.txt
pyyaml
requests
pyopenssl
pyzabbix
